dos: add OpenRelative
