+++ rom/dos/dos.conf
@@ -257,4 +257,5 @@ LONG GetSegListInfo(BPTR seglist, const struct TagItem *taglist) (D0, A0)
 BOOL AssignAddToList(CONST_STRPTR name, BPTR lock, ULONG position) (D1, D2, D3)
 # AROS-specific
 BPTR LockRelative(BPTR lock, CONST_STRPTR name, LONG accessMode) (D1,D2,D3)
+BPTR OpenRelative(BPTR lock, CONST_STRPTR name, LONG accessMode) (D1,D2,D3)
 ##end functionlist
