+++ rom/dos/mmakefile.src
@@ -53,7 +53,7 @@ FUNCTIONS := abortpkt addbuffers adddosentry addpart addsegment \
 	     isinteractive loadseg lock lockdoslist lockrelative lockrecord lockrecords \
 	     makedosentry makelink matchend matchfirst matchnext matchpattern \
 	     matchpatternnocase maxcli namefromfh namefromlock newloadseg nextdosentry nil \
-	     open openfromlock output parentdir parentoffh parsepattern \
+	     open openfromlock openrelative output parentdir parentoffh parsepattern \
 	     parsepatternnocase pathpart printfault putstr read readargs \
 	     readitem relabel readlink remassignlist remdosentry remsegment rename \
 	     replypkt runcommand samedevice samelock scanvars seek \
